Fix authenticator, minor styling changes.
