+++ src/App.css
@@ -42,7 +42,8 @@
   right: 0;
   overflow: hidden;
   height: fit-content;
-  border: 1px solid black;
+  border-bottom: 1px solid black;
+  border-left: 1px solid black;
 }
 
 .menuNav button {
